Add tooling to list hardforks upgrades
